staging: fsl-dpaa2/eth: Add XDP redirect support
Implement support for the XDP_REDIRECT action.

The redirected frame is transmitted and confirmed on the regular
Tx/Tx conf queues. Frame is marked with the "XDP" type in the
software annotation, since it requires special treatment.

In case of an error on transmit, Rx interface discards the frame
and performs all necessary cleanup.

We don't have good hardware support for flushing, so ndo_xdp_flush
doesn't do anything for now; ndo_xdp_xmit performs the actual Tx
operation on the spot.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@nxp.com>