driver: mtd: fsl_qspi: Remap QSPI address space for LS1088 and LS2088
QSPI support large memories (up to 4GB each) and this memory may
be accessed during the SoC reset sequence for loading RCW or the boot code.
Since this occurs before the system MMUs are configured, so QSPI memory
must be available in the lower 4GB space of the SoC address map.
Therefore, QSPI memory regions were split into two parts: a 256MB region that
is located in the least significant 4GB of the SoC address space and
a 3.75GB region that is located above the least significant 4GB of the
SoC address space.

QSPI Address Remapping
SoC Address 			QSPI Address
0x00_2000_0000 - 0x00_2FFF_FFFF	0x00_0000_0000 - 0x00_0FFF_FFFF  First 256MB
0x04_1000_0000 - 0x04_FFFF_FFFF 0x00_1000_0000 - 0x00_FFFF_FFFF  Last 3.75GB

Signed-off-by: Suresh Gupta <suresh.gupta@nxp.com>