crypto: caam - initialize RNG in deterministic (test) mode
Palladium and CFP emulators do not correctly model the ring oscillator
used by CAAM RNG as the entropy source for its TRNG.
This causes the statistical tests to fail.

As a workaround, initialize RNG in deterministic (test) mode.
In this mode, TRNG (and thus the ring oscillator) is not used
for seeding the DRNG.

Signed-off-by: Horia Geantă <horia.geanta@nxp.com>