drm/mali-dp: Add writeback support for DP500.
Mali DP500 behaves differently from the rest of the Mali DP IP,
in that it does not have a one-shot mode and keeps writing the
content of the current frame to the provided memory area until
stopped. As a way of emulating the one-shot behaviour, we are
going to use the CVAL interrupt that is being raised at the
start of each frame, during prefetch phase, to act as End-of-Write
signal, but with a twist: we are going to disable the memory
write engine right after it has been enabled, because the bit
controlling the enabling will only be acted upon on the next
vblank/prefetch.

CVAL interrupt will fire every time the global CVAL bit gets set
during the next prefetch phase, so we need a state byte to track
the memory write enabling.

Signed-off-by: Liviu Dudau <liviu.dudau@arm.com>