enetc: Call register_netdev as late as possible, refactoring
This removes the risk of a race at probe, once the netdevice gets
registered. This is a requirement that the upstream maintainers
always verify during review, and point it out if not implemented.
Note that request_irq must be moved to open(), to be able to call
it after register_netdev().

Signed-off-by: Claudiu Manoil <claudiu.manoil@nxp.com>