felix: convert to use interrupt for injection timestamping
Convert to use interrupt for injection timestamping, instead
of polling before. The INTB interrupt is not only for PTP
timestamp ready event, but also for TSN preemption event.
So handling preemption is still a TODO thing in ISR, otherwise
there will be interrupt storm if it's triggered without cleaning
interrupt status in ISR.

Signed-off-by: Alex Marginean <alexandru.marginean@nxp.com>
Signed-off-by: Yangbo Lu <yangbo.lu@nxp.com>