arm64: dts: ls1028a: disable eno0 and eno1 for the FII LS1028A-TSN board
These ports are not used on this board.

In mainline, all Ethernet ports are disabled in the ls1028a.dtsi, and it
is a per-board decision to enable them. To keep the delta smaller, we
choose to do the opposite here (i.e. simply disable them in the board
DTS).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>