net: dsa: felix: replace 802.1CB stream mact_idx with mac and vid
Let's take the function below.

static void ocelot_prove_mac_table_entries_can_move(struct ocelot *ocelot)
{
	unsigned char mac1[ETH_ALEN] = {0x00, 0x04, 0x9f, 0x63, 0x35, 0xea};
	unsigned char mac2[ETH_ALEN] = {0x00, 0x04, 0x9f, 0x63, 0x35, 0xeb};
	int row, bucket, arbitrary_pgid = 4;
	int vid1 = 102;
	int vid2 = 103;
	int err;

	err = ocelot_mact_learn(ocelot, arbitrary_pgid, mac1, vid1,
				ENTRYTYPE_LOCKED);
	if (err)
		return;

	err = ocelot_mact_lookup(ocelot, mac1, vid1, &row, &bucket);
	if (err)
		return;

	dev_info(ocelot->dev,
		 "Address 1 (mac %pM vid %d) is in MAC table row %d bucket %d\n",
		 mac1, vid1, row, bucket);

	err = ocelot_mact_learn(ocelot, arbitrary_pgid, mac2, vid2,
				ENTRYTYPE_LOCKED);
	if (err)
		return;

	err = ocelot_mact_lookup(ocelot, mac2, vid2, &row, &bucket);
	if (err)
		return;

	dev_info(ocelot->dev,
		 "Address 2 (mac %pM vid %d) is in MAC table row %d bucket %d\n",
		 mac2, vid2, row, bucket);

	err = ocelot_mact_lookup(ocelot, mac1, vid1, &row, &bucket);
	if (err)
		return;

	dev_info(ocelot->dev,
		 "Address 1 (mac %pM vid %d) is in MAC table row %d bucket %d\n",
		 mac1, vid1, row, bucket);
}

What will it print?

Address 1 (mac 00:04:9f:63:35:ea vid 102) is in MAC table row 917 bucket 0
Address 2 (mac 00:04:9f:63:35:eb vid 103) is in MAC table row 917 bucket 0
Address 1 (mac 00:04:9f:63:35:ea vid 102) is in MAC table row 917 bucket 1

What does this mean?

The ROW portion of a FDB entry's position within the MAC table is
statically determined using an 11-bit hash derived from the {DMAC, VID}
key. Within a row, there can be up to 4 buckets, each bucket holding 1
MAC table entry.

But when the hashes of 2 addresses collide and they end up in the same
row (as in the above example, with address 1 = "mac 00:04:9f:63:35:ea
vid 102" and address 2 = "mac 00:04:9f:63:35:eb vid 103"), things don't
happen quite as you might expect. Namely, the second address appears to
be installed by the switch at the same row and bucket as the first
address. So is the first address overwritten? No, it has been moved by
the switch, automatically, to bucket 1.

Let's leave aside for a moment the fact that it's stupid of the hardware
to do that. But the current 802.1CB code has a problem. It makes the
fundamental assumption that the stream IDs have a constant index within
the MAC table, assumption which we proved is false.

To be precise, when a 'tsntool cbstreamidset' with address 2 is run,
it will create in the kernel two streamid structures which both have the
same MAC table index. Then, when one of these streams is configured for
splitting with a command such as 'tsntool cbgen', it will in fact affect
both streams, by calling streamid_multi_forward_set().

The effect of this bug is that, when multiple 802.1CB stream ID rules
are used, stream splitting works with some MAC addresses, but not with
others.

Fix it by making the 802.1CB memorize the MAC and VID instead of the
static MAC table index. The MAC table index will always be looked up by
the driver when something needs to be done with the streams.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>