net: stmmac: fix taprio schedule configuration
When configuring a 802.1Qbv schedule through the tc taprio qdisc on an NXP
i.MX8MPlus device, the effective cycle time differed from the requested one
by N*96ns, with N number of entries in the Qbv Gate Control List. This is
because the driver was adding a 96ns margin to each interval of the GCL,
apparently to account for the IPG. The problem was observed on NXP
i.MX8MPlus devices but likely affected all devices relying on the same
configuration callback (dwmac 4.00, 4.10, 5.10 variants).

Fix the issue by removing the margins, and simply set up the MAC with the
provided cycle time value. This is the behavior expected by the user-space
API and of various other Ethernet MAC supporting taprio, including the
dwxgmac variant of stmmac.

Signed-off-by: Yannick Vignon <yannick.vignon@nxp.com>