net: dsa: felix: disable ingress drop mode on ports
The ocelot switches do not support flow control on the NPI interface
(CPU port in DSA terms). And since the default CPU port is the 2.5Gbps
swp4 (connected to eno2), this means that for certain traffic streams
such as iperf TCP termination, the switch must forward a 2.5Gbps stream
between the NPI port and external interfaces running at 1Gbps. The CPU
doesn't know that it needs to stop sending, and when the switch buffers
fill, it drops extra TCP segments in a bursty fashion, causing the TCP
congestion control to go haywire and stabilize at a very low size,
giving us bad throughput (300 - 500 Mbps).

This is a very large issue that in itself cannot be addressed by a
one-line commit, however there is a desire to have a little bit better
performance for TCP termination.

This patch disables the controlling watermarks and lets the tail
dropping threshold kick in. In turn, this allows any single stream in
the switch to use up the entire frame buffer memory. Because we let the
switch buffer more packets before dropping now, this moves the issue a
little bit later, and a single iperf3 TCP stream is able to achieve
1Gbps line rate.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>